Initialize Checkable.dim in the constructor rather than later.
Only attempt to verify a checksum if the database tells us it exists.

If an expected checksum is not found, consult Checkable.addable() to
decide whether to add the file to the sample.

Ensure dimension objects are singletons by name (i.e., there's only
one 'cos' Dimension, for example).

Avoid keeping the database open in Dimension.