When we persist the Checkable at the end of check(), log the fact along with the values of fails and last_check.
In Checkable.verify(), don't give up on timeouts unless before is
empty. If we're no longer getting anything from hsi, give up, but if
we're still getting non-error verbiage, keep waiting and expecting for
a prompt or an error.

In HSI.quit(), don't generate an exception on a timeout. Just catch
it, close the child connection, and quit.